internal/plugins/helm: refactor for organization and isolation
