+++ internal/plugins/helm/v1/scaffolds/internal/templates/config/crd/crd.go
