+++ internal/plugins/helm/v1/scaffolds/internal/templates/config/manager/kustomization.go
