+++ internal/plugins/helm/v1/scaffolds/internal/templates/config/prometheus/kustomization.go
