+++ internal/plugins/helm/v1/scaffolds/internal/templates/config/rbac/auth_proxy_role_binding.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package templates
+package rbac
 
 import (
 	"path/filepath"
