+++ internal/plugins/helm/v1/scaffolds/internal/templates/config/rbac/manager_role.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package templates
+package rbac
 
 import (
 	"bytes"
