+++ internal/plugins/helm/v1/scaffolds/internal/templates/config/samples/crd_sample.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package templates
+package samples
 
 import (
 	"fmt"
