+++ internal/plugins/helm/v1/scaffolds/internal/templates/dockerfile.go
