+++ internal/plugins/helm/v1/scaffolds/internal/templates/gitignore.go
