+++ internal/plugins/helm/v1/scaffolds/internal/templates/watches.go
